refactor(charts): rename charts to app-charts
