+++ src/pages/stats/index.vue
@@ -9,7 +9,7 @@
       :dutch-data="dutchData"
     />
 
-    <charts
+    <app-charts
       v-if="selectedData && selectedDutchData"
       :selected-data="selectedData"
       :dutch-data="selectedDutchData"
@@ -17,30 +17,28 @@
       title="Charts toggle"
     />
 
-    <nuxt-link to="/" class="button button--primary" >Opnieuw beginnen</nuxt-link>
-    <nuxt-link
-      to="/shop"
-      class="button button--secondary"
+    <nuxt-link to="/" class="button button--primary"
+      >Opnieuw beginnen</nuxt-link
     >
+    <nuxt-link to="/shop" class="button button--secondary">
       Naar webshop
     </nuxt-link>
   </main>
 </template>
 
 <script>
-import { mapGetters } from 'vuex'
-import { SET_CARECOMPANIES, SET_DUTCH_DATA } from '~/store/mutation-types'
-import dutchDataJson from '~/static/data/dutch-stats'
-import CompanySearch from '~/components/company-search/company-search'
-import YearSelection from '~/components/year-selection/year-selection'
-import Charts from '~/components/charts/charts'
-
+import { mapGetters } from "vuex"
+import { SET_CARECOMPANIES, SET_DUTCH_DATA } from "~/store/mutation-types"
+import dutchDataJson from "~/static/data/dutch-stats"
+import CompanySearch from "~/components/company-search/company-search"
+import YearSelection from "~/components/year-selection/year-selection"
+import AppCharts from "~/components/app-charts/app-charts"
 
 export default {
   components: {
     CompanySearch,
     YearSelection,
-    Charts
+    AppCharts
   },
   data() {
     return {
@@ -49,32 +47,34 @@ export default {
   },
   computed: {
     ...mapGetters({
-      careCompanies: 'careCompanies',
-      selectedCareCompany: 'selectedCareCompany',
-      selectedData: 'selectedData',
-      selectedDutchData: 'selectedDutchData'
+      careCompanies: "careCompanies",
+      selectedCareCompany: "selectedCareCompany",
+      selectedData: "selectedData",
+      selectedDutchData: "selectedDutchData"
     }),
     async rawCompanies() {
-      const module = await import('~/static/data/pointer-raw')
+      const module = await import("~/static/data/pointer-raw")
       return module.default
-    },
+    }
   },
   async mounted() {
     if (!this.careCompanies) {
-      this.$store.commit(SET_CARECOMPANIES, { careCompanies: await this.rawCompanies })
+      this.$store.commit(SET_CARECOMPANIES, {
+        careCompanies: await this.rawCompanies
+      })
     }
   }
 }
 </script>
 
 <style lang="scss">
-.stats-page{
+.stats-page {
   padding: $spacing-small $spacing-medium;
 }
 
-.stats-page_title{
-first-letter {text-transform: uppercase};
+.stats-page_title {
+  // first-letter {
+  //   text-transform: uppercase;
+  // }
 }
-
-
 </style>
